fixed import and moved the util function
